+++ src/PIL/XpmImagePlugin.py
@@ -18,7 +18,7 @@
 import re
 
 from . import Image, ImageFile, ImagePalette
-from ._binary import i8, o8
+from ._binary import o8
 
 # XPM header
 xpm_head = re.compile(b'"([0-9]*) ([0-9]*) ([0-9]*) ([0-9]*)')
@@ -72,7 +72,7 @@ class XpmImageFile(ImageFile.ImageFile):
             elif s[-1:] in b"\r\n":
                 s = s[:-1]
 
-            c = i8(s[1])
+            c = s[1]
             s = s[2:-2].split()
 
             for i in range(0, len(s), 2):
